fix(se): correct sonarlint 2.0

+++ src/ekf_state_est/src/kalman_filter/observation_models.cpp
@@ -108,17 +108,17 @@ Eigen::MatrixXf ObservationModel::get_state_to_observation_matrix(
   // where x is the x coordinate of the robot, y is the y coordinate of the robot, theta is the
   // orientation of the robot, m is the x coordinate of the landmark, n is the y coordinate of the
   // landmark
-  low_jacobian(0, 0) = (float) -cos(-expected_state(2));
-  low_jacobian(0, 1) = (float) sin(-expected_state(2));
-  low_jacobian(0, 2) = (float) -expected_state(landmark_index) * sin(expected_state(2)) +
+  low_jacobian(0, 0) = -cos(-expected_state(2));
+  low_jacobian(0, 1) = sin(-expected_state(2));
+  low_jacobian(0, 2) = -expected_state(landmark_index) * sin(expected_state(2)) +
                        expected_state(landmark_index + 1) * cos(-expected_state(2)) +
                        expected_state(0) * sin(expected_state(2)) -
                        expected_state(1) * cos(-expected_state(2));
   low_jacobian(0, 3) = 0;
   low_jacobian(0, 4) = 0;
   low_jacobian(0, 5) = 0;
-  low_jacobian(0, 6) = (float) cos(-expected_state(2));
-  low_jacobian(0, 7) = (float) -sin(-expected_state(2));
+  low_jacobian(0, 6) = cos(-expected_state(2));
+  low_jacobian(0, 7) = -sin(-expected_state(2));
 
   low_jacobian(1, 0) = (float) -sin(-expected_state(2));
   low_jacobian(1, 1) = (float) -cos(-expected_state(2));
